Refactors Faraday::Connection to fix Rubocop ClassLength cop
